requirements.txt
Flask==3.0.0
Pillow==11.0.0